+++ scripts/forge_couple.py
@@ -1,4 +1,5 @@
 from modules import scripts
+from json import dumps
 import re
 
 from scripts.couple_mapping import (
@@ -7,12 +8,12 @@ from scripts.couple_mapping import (
     advanced_mapping,
     mask_mapping,
 )
-from scripts.couple_ui import couple_UI, validata_mapping, parse_mapping, hook_component
+from scripts.couple_ui import couple_UI, validate_mapping, parse_mapping, hook_component
 
 from scripts.attention_couple import AttentionCouple
 forgeAttentionCouple = AttentionCouple()
 
-VERSION = "1.3.7"
+VERSION = "1.4.3"
 
 
 class ForgeCouple(scripts.Script):
@@ -86,7 +87,7 @@ class ForgeCouple(scripts.Script):
             self.couples = None
             return
 
-        if (mode == "Advanced") and not validata_mapping(mapping):
+        if (mode == "Advanced") and not validate_mapping(mapping):
             self.couples = None
             return
 
@@ -96,6 +97,20 @@ class ForgeCouple(scripts.Script):
             )
             self.couples = None
             return
+
+        # ===== Infotext =====
+        p.extra_generation_params["forge_couple"] = True
+        p.extra_generation_params["forge_couple_separator"] = (
+            "\n" if not separator.strip() else separator.strip()
+        )
+        p.extra_generation_params["forge_couple_mode"] = mode
+        if mode == "Basic":
+            p.extra_generation_params["forge_couple_direction"] = direction
+            p.extra_generation_params["forge_couple_background"] = background
+            p.extra_generation_params["forge_couple_background_weight"] = background_weight
+        elif mode == "Advanced":
+            p.extra_generation_params["forge_couple_mapping"] = dumps(mapping)
+        # ===== Infotext =====
 
         self.couples = couples
 
@@ -115,20 +130,6 @@ class ForgeCouple(scripts.Script):
 
         if not enable or not self.couples:
             return
-
-        # ===== Infotext =====
-        p.extra_generation_params["forge_couple"] = True
-        p.extra_generation_params["forge_couple_separator"] = (
-            "\n" if not separator.strip() else separator.strip()
-        )
-        p.extra_generation_params["forge_couple_mode"] = mode
-        p.extra_generation_params["forge_couple_background_weight"] = background_weight
-        if mode == "Basic":
-            p.extra_generation_params["forge_couple_direction"] = direction
-            p.extra_generation_params["forge_couple_background"] = background
-        elif mode == "Advanced":
-            p.extra_generation_params["forge_couple_mapping"] = mapping
-        # ===== Infotext =====
 
         # ===== Init =====
         unet = p.sd_model.forge_objects.unet